change environment candidates by test cases ids
